Add reencryption

+++ crypt.go
@@ -492,7 +492,14 @@ func ActivateVolumeWithKeyData(volumeName, sourceDevicePath string, authRequesto
 				continue
 			}
 
-			candidates = append(candidates, &keyCandidate{KeyData: kd, slot: token.Keyslots()[0]})
+			var slot int
+			if len(token.Keyslots()) == 1 {
+				slot = token.Keyslots()[0]
+			} else {
+				// Two keyslots, typically when reencryption is in progress
+				slot = luks2.AnySlot
+			}
+			candidates = append(candidates, &keyCandidate{KeyData: kd, slot: slot})
 		}
 	}
 
@@ -659,8 +666,8 @@ func InitializeLUKS2Container(devicePath, label string, key DiskUnlockKey, optio
 
 	token := luksview.KeyDataToken{
 		TokenBase: luksview.TokenBase{
-			TokenKeyslot: 0,
-			TokenName:    initialKeyslotName}}
+			TokenKeyslots: []int{0},
+			TokenName:     initialKeyslotName}}
 	if err := luks2ImportToken(devicePath, &token, nil); err != nil {
 		return xerrors.Errorf("cannot import token: %w", err)
 	}
@@ -735,8 +742,8 @@ func addLUKS2ContainerKey(devicePath, keyslotName string, existingKey, newKey Di
 	// a single atomic transaction ¯\_(ツ)_/¯
 
 	tokenBase := luksview.TokenBase{
-		TokenName:    keyslotName,
-		TokenKeyslot: freeSlot}
+		TokenName:     keyslotName,
+		TokenKeyslots: []int{freeSlot}}
 	if err := luks2ImportToken(devicePath, newToken(&tokenBase), nil); err != nil {
 		return xerrors.Errorf("cannot import token: %w", err)
 	}
@@ -887,6 +894,10 @@ func DeleteLUKS2ContainerKey(devicePath, keyslotName string) error {
 
 	removeOrphanedTokens(devicePath, view)
 
+	if len(token.Keyslots()) != 1 {
+		return fmt.Errorf("token has %v keyslot(s)", len(token.Keyslots()))
+	}
+
 	slot := token.Keyslots()[0]
 	if err := luks2KillSlot(devicePath, slot); err != nil {
 		return xerrors.Errorf("cannot kill existing slot %d: %w", slot, err)
@@ -937,15 +948,15 @@ func renameLUKS2ContainerKey(nonAtomic *nonAtomicOperationAllowedFlag, devicePat
 	case *luksview.KeyDataToken:
 		newToken = &luksview.KeyDataToken{
 			TokenBase: luksview.TokenBase{
-				TokenKeyslot: t.TokenKeyslot,
-				TokenName:    newName},
+				TokenKeyslots: t.TokenKeyslots,
+				TokenName:     newName},
 			Priority: t.Priority,
 			Data:     t.Data}
 	case *luksview.RecoveryToken:
 		newToken = &luksview.RecoveryToken{
 			TokenBase: luksview.TokenBase{
-				TokenKeyslot: t.TokenKeyslot,
-				TokenName:    newName}}
+				TokenKeyslots: t.TokenKeyslots,
+				TokenName:     newName}}
 	default:
 		return errors.New("cannot rename key with unexpected token type")
 	}
@@ -1032,8 +1043,8 @@ func NameLegacyLUKS2ContainerKey(devicePath string, keyslot int, newName string)
 
 	token := &luksview.RecoveryToken{
 		TokenBase: luksview.TokenBase{
-			TokenName:    newName,
-			TokenKeyslot: keyslot,
+			TokenName:     newName,
+			TokenKeyslots: []int{keyslot},
 		},
 	}
